Extract value calculation to helper file
