+++ src/entities/addLiquidity/helpers.ts
@@ -1,4 +1,3 @@
-import { NATIVE_ASSETS } from '@/utils';
 import { AddLiquidityAmounts } from '../types';
 import { AddLiquidityBaseBuildCallInput, AddLiquidityKind } from './types';
 
@@ -25,15 +24,4 @@ export const getAmountsCall = (
             };
         }
     }
-};
-
-export const getValue = (input: AddLiquidityBaseBuildCallInput): bigint => {
-    let value = 0n;
-    if (input.wethIsEth) {
-        value =
-            input.amountsIn.find(
-                (a) => a.token.address === NATIVE_ASSETS[input.chainId].wrapped,
-            )?.amount ?? 0n;
-    }
-    return value;
 };
