+++ src/entities/initPool/initPoolV2/composableStable/initPoolComposableStable.ts
@@ -12,8 +12,10 @@ import {
     InitPoolInputV2,
 } from '../../types';
 import { vaultV2Abi } from '../../../../abi';
-import { VAULT, MAX_UINT256, ZERO_ADDRESS } from '../../../../utils';
+import { VAULT, MAX_UINT256 } from '../../../../utils';
 import { Token } from '@/entities/token';
+import { getValue } from '@/entities/utils/getValue';
+import { TokenAmount } from '@/entities/tokenAmount';
 
 export class InitPoolComposableStable implements InitPoolBase {
     buildCall(
@@ -33,6 +35,7 @@ export class InitPoolComposableStable implements InitPoolBase {
             maxAmountsIn: amounts.maxAmountsIn,
             userData,
             fromInternalBalance: input.fromInternalBalance ?? false,
+            wethIsEth: !!input.wethIsEth,
         });
         const call = encodeFunctionData({
             abi: vaultV2Abi,
@@ -40,14 +43,15 @@ export class InitPoolComposableStable implements InitPoolBase {
             args,
         });
 
-        const value = input.amountsIn.find(
-            (a) => a.address === ZERO_ADDRESS,
-        )?.rawAmount;
+        const amountsIn = input.amountsIn.map((a) => {
+            const token = new Token(input.chainId, a.address, a.decimals);
+            return TokenAmount.fromRawAmount(token, a.rawAmount);
+        });
 
         return {
             call,
             to: VAULT[input.chainId] as Address,
-            value: value === undefined ? 0n : value,
+            value: getValue(amountsIn, !!input.wethIsEth),
         };
     }
